requirements.txt
# ActiveMQ
stomp.py==8.2.0

# Base de données
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API et Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Utilitaires
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0 